--- core_java/jbdc/sql/database.sql
@@ -1,2 +1,0 @@
-CREATE DATABASE hplussport;
-GRANT ALL PRIVILEGES ON DATABASE hplussport TO postgres;
