--- core_java/jbdc/src/main/java/ca/jrvs/apps/jdbc/util/DataTransferObject.java
@@ -1,7 +1,0 @@
-package ca.jrvs.apps.jdbc.util;
-
-public interface DataTransferObject {
-
-  long getId();
-}
-
