+++ core_java/jdbc/assets/ERD.PNG
