+++ core_java/jdbc/sql/customer.sql
