+++ core_java/jdbc/sql/database.sql
@@ -1,0 +1,2 @@
+CREATE DATABASE hplussport;
+GRANT ALL PRIVILEGES ON DATABASE hplussport TO postgres;
