+++ core_java/jdbc/src/main/java/ca/jrvs/apps/jdbc/util/DataTransferObject.java
@@ -1,0 +1,7 @@
+package ca.jrvs.apps.jdbc.util;
+
+public interface DataTransferObject {
+
+  long getId();
+}
+
